Rename from "proof.c" to "theorem.c" because it defines theorems.
