+++ build.sh
@@ -1,6 +1,6 @@
 #!/bin/sh -ex
 
 CC="gcc -ansi -pedantic"
-for MODULE in assert formula assumptions proof; do
+for MODULE in assert formula assumptions theorem; do
     (cd src && ${CC} ${CFLAGS} -I../include -c $MODULE.c -o $MODULE.o)
 done
